requirements.txt
Flask==3.1.1
elevenlabs==1.51.0
python-dotenv
